chore: rename `new` to `future`
