+++ src/browser/setup-worker.ts
@@ -1,11 +1,11 @@
 import { invariant } from 'outvariant'
 import { isNodeProcess } from 'is-node-process'
 import { WebSocketInterceptor } from '@mswjs/interceptors/WebSocket'
-import { defineNetwork, NetworkHandlersApi } from '#core/new/define-network'
-import { type AnyHandler } from '#core/new/handlers-controller'
-import { InterceptorSource } from '#core/new/sources/interceptor-source'
+import { defineNetwork, NetworkHandlersApi } from '#core/future/define-network'
+import { type AnyHandler } from '#core/future/handlers-controller'
+import { InterceptorSource } from '#core/future/sources/interceptor-source'
 import { type UnhandledRequestStrategy } from '#core/utils/request/onUnhandledRequest'
-import { fromLegacyOnUnhandledRequest } from '#core/new/compat'
+import { fromLegacyOnUnhandledRequest } from '#core/future/compat'
 import { devUtils } from '#core/utils/internal/devUtils'
 import { supportsServiceWorker } from './utils/supports'
 import { ServiceWorkerSource } from './sources/service-worker-source'
