+++ src/core/future/define-network.ts
