+++ src/node/setup-server-common.ts
@@ -1,10 +1,10 @@
 import type { PartialDeep } from 'type-fest'
 import { Interceptor } from '@mswjs/interceptors'
-import { type NetworkApi, defineNetwork } from '#core/new/define-network'
-import { type AnyHandler } from '#core/new/handlers-controller'
-import { HandlersController } from '#core/new/handlers-controller'
-import { InterceptorSource } from '#core/new/sources/interceptor-source'
-import { fromLegacyOnUnhandledRequest } from '#core/new/compat'
+import { type NetworkApi, defineNetwork } from '#core/future/define-network'
+import { type AnyHandler } from '#core/future/handlers-controller'
+import { HandlersController } from '#core/future/handlers-controller'
+import { InterceptorSource } from '#core/future/sources/interceptor-source'
+import { fromLegacyOnUnhandledRequest } from '#core/future/compat'
 import type { ListenOptions, SetupServerCommon } from './glossary'
 
 export function createSetupServerCommonApi(
